+++ picoquic/util.h
@@ -45,8 +45,8 @@ void debug_printf_suspend(void);
 void debug_printf_resume(void);
 
 extern const picoquic_connection_id_t picoquic_null_connection_id;
-size_t picoquic_format_connection_id(uint8_t* bytes, size_t bytes_max, picoquic_connection_id_t cnx_id);
-size_t picoquic_parse_connection_id(uint8_t* bytes, uint8_t len, picoquic_connection_id_t *cnx_id);
+uint32_t picoquic_format_connection_id(uint8_t* bytes, size_t bytes_max, picoquic_connection_id_t cnx_id);
+uint32_t picoquic_parse_connection_id(uint8_t* bytes, uint8_t len, picoquic_connection_id_t *cnx_id);
 int picoquic_is_connection_id_null(picoquic_connection_id_t cnx_id);
 int picoquic_compare_connection_id(picoquic_connection_id_t * cnx_id1, picoquic_connection_id_t * cnx_id2);
 uint64_t picoquic_val64_connection_id(picoquic_connection_id_t cnx_id);
